--- examples/cloud-grpc-security-sleuth-zipkin/cloud-grpc-security-sleuth-zipkin-order/src/main/java/my/suveng/istiocloudorderdemo/controller/Controller.java
@@ -1,23 +1,0 @@
-package my.suveng.istiocloudorderdemo.controller;
-
-import my.suveng.istiocloudorderdemo.consumer.Service;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- *
- * @author suwenguang
- **/
-@RestController
-public class Controller {
-
-	@Autowired
-	Service service;
-
-	@RequestMapping("/pay")
-	public String pay(){
-		service.t1();
-		return "success";
-	}
-}
