+++ examples/cloud-grpc-security-sleuth-zipkin/cloud-grpc-security-sleuth-zipkin-order/src/test/java/my/suveng/istiocloudorderdemo/CloudOrderDemoApplicationTests.java
@@ -1,0 +1,13 @@
+package my.suveng.istiocloudorderdemo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CloudOrderDemoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
