--- examples/cloud-grpc-security-sleuth-zipkin/cloud-grpc-security-sleuth-zipkin-order/src/test/java/my/suveng/istiocloudorderdemo/IstioCloudOrderDemoApplicationTests.java
@@ -1,13 +1,0 @@
-package my.suveng.istiocloudorderdemo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class IstioCloudOrderDemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
