+++ examples/cloud-grpc-security-sleuth-zipkin/cloud-grpc-security-sleuth-zipkin-pay/src/test/java/my/suveng/cloud_pay_demo/CloudPayDemoApplicationTests.java
@@ -1,0 +1,13 @@
+package my.suveng.cloud_pay_demo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CloudPayDemoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
