--- examples/cloud-grpc-security-sleuth-zipkin/cloud-grpc-security-sleuth-zipkin-pay/src/test/java/my/suveng/istiocloudpaydemo/IstioCloudPayDemoApplicationTests.java
@@ -1,13 +1,0 @@
-package my.suveng.istiocloudpaydemo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class IstioCloudPayDemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
