--- examples/cloud-grpc-security-sleuth-zipkin/cloud-grpc-security-sleuth-zipkin-user/src/main/java/my/suveng/istioclouduserdemo/config/ThreadConfig.java
@@ -1,36 +1,0 @@
-package my.suveng.istioclouduserdemo.config;
-
-import org.springframework.beans.factory.BeanFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cloud.sleuth.instrument.async.LazyTraceExecutor;
-import org.springframework.cloud.sleuth.instrument.async.LazyTraceThreadPoolTaskExecutor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.AsyncConfigurerSupport;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-/**
- * 异步线程池配置
- * 使用
- * @author suwenguang
- **/
-@Configuration
-@EnableAsync
-public class ThreadConfig extends AsyncConfigurerSupport {
-	@Autowired
-	private BeanFactory beanFactory;
-
-	@Override
-	public Executor getAsyncExecutor() {
-		ThreadPoolTaskExecutor threadPoolTaskExecutor = new ThreadPoolTaskExecutor();
-		threadPoolTaskExecutor.setCorePoolSize(7);
-		threadPoolTaskExecutor.setMaxPoolSize(42);
-		threadPoolTaskExecutor.setQueueCapacity(11);
-		threadPoolTaskExecutor.setThreadNamePrefix("async-");
-		threadPoolTaskExecutor.initialize();
-
-		return new LazyTraceThreadPoolTaskExecutor(beanFactory, threadPoolTaskExecutor);
-	}
-}
