+++ examples/cloud-grpc-security-sleuth-zipkin/cloud-grpc-security-sleuth-zipkin-user/src/test/java/my/suveng/cloud_user_demo/CloudUserDemoApplicationTests.java
@@ -1,0 +1,13 @@
+package my.suveng.cloud_user_demo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CloudUserDemoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
