--- examples/cloud-grpc-security-sleuth-zipkin/cloud-grpc-security-sleuth-zipkin-user/src/test/java/my/suveng/istioclouduserdemo/IstioCloudUserDemoApplicationTests.java
@@ -1,13 +1,0 @@
-package my.suveng.istioclouduserdemo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class IstioCloudUserDemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
